# specs/001-i-want-to/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Lift Tracker API
  description: REST API for workout tracking application
  version: 1.0.0
  contact:
    name: Lift Tracker Team
servers:
  - url: https://api.lifttracker.com/v1
    description: Production server
  - url: https://localhost:7001/api
    description: Development server
security:
  - GoogleOAuth: []
paths:
...
  /workout-sessions/{sessionId}:
    get:
      tags: [Workout Sessions]
      summary: Get specific workout session with exercises
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workout session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Workout Sessions]
      summary: Update workout session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutSession'
      responses:
        '200':
          description: Workout session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Workout Sessions]
      summary: Delete workout session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Workout session deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Strength Lift endpoints
...
components:
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Access to email address
            profile: Access to profile information
  schemas:
...
    WorkoutSession:
      type: object
      required: [id, userId, date, createdDate, modifiedDate]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        notes:
          type: string
          maxLength: 1000
        createdDate:
          type: string
          format: date-time
        modifiedDate:
          type: string
          format: date-time
    WorkoutSessionDetail:
      allOf:
        - $ref: '#/components/schemas/WorkoutSession'
        - type: object
          properties:
            strengthLifts:
              type: array
              items:
                $ref: '#/components/schemas/StrengthLift'
            metconWorkouts:
              type: array
              items:
                $ref: '#/components/schemas/MetconWorkout'
...
    UpdateWorkoutSession:
      type: object
      properties:
        notes:
          type: string
          maxLength: 1000
    ExerciseType:
      type: object
      required: [id, name, category, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 100
        category:
          type: string
          maxLength: 50
        isActive:
          type: boolean
    StrengthLift:
      type: object
      required: [id, workoutSessionId, exerciseTypeId, setStructure, weight, order]
      properties:
        id:
          type: string
          format: uuid
        workoutSessionId:
          type: string
          format: uuid
        exerciseTypeId:
          type: integer
        exerciseType:
          $ref: '#/components/schemas/ExerciseType'
        setStructure:
          type: string
          enum: [SetsReps, EMOM, AMRAP, TimeBased]
        sets:
          type: integer
          minimum: 1
          maximum: 50
        reps:
          type: integer
          minimum: 1
          maximum: 500
        weight:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        additionalWeight:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        duration:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        restPeriod:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        comments:
          type: string
          maxLength: 500
        order:
          type: integer
          minimum: 1
...
    MetconType:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 50
        description:
          type: string
          maxLength: 200
        isActive:
          type: boolean
    MovementType:
      type: object
      required: [id, name, category, measurementType, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 100
        category:
          type: string
          maxLength: 50
        measurementType:
          type: string
          enum: [Reps, Distance]
        isActive:
          type: boolean
    MetconWorkout:
      type: object
      required: [id, workoutSessionId, metconTypeId, order]
      properties:
        id:
          type: string
          format: uuid
        workoutSessionId:
          type: string
          format: uuid
        metconTypeId:
          type: integer
        metconType:
          $ref: '#/components/schemas/MetconType'
        totalTime:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        roundsCompleted:
          type: integer
          minimum: 0
          maximum: 1000
        notes:
          type: string
          maxLength: 1000
        order:
          type: integer
          minimum: 1
        movements:
          type: array
          items:
            $ref: '#/components/schemas/MetconMovement'
    MetconMovement:
      type: object
      required: [id, metconWorkoutId, movementTypeId, order]
      properties:
        id:
          type: string
          format: uuid
        metconWorkoutId:
          type: string
          format: uuid
        movementTypeId:
          type: integer
        movementType:
          $ref: '#/components/schemas/MovementType'
        reps:
          type: integer
          minimum: 1
          maximum: 10000
        distance:
          type: number
          format: decimal
          minimum: 0
        weight:
          type: number
          format: decimal
          minimum: 0
          multipleOf: 0.25
        order:
          type: integer
          minimum: 1
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'